EDU-1303 Remove no longer needed base class
